add electron script

+++ scripts/batchRemoveDist.mjs
@@ -57,7 +57,33 @@ function deleteFolder(path) {
 }
 
 const start = () => {
+  const list = [
+    {
+      type: 'list',
+      message: '选择要操作的文件夹:',
+      name: 'OperationType',
+      choices: [
+        {
+          name: 'dist',
+        },
+        {
+          name: 'dist-electron',
+        },
+      ],
+    },
+  ]
+  inquirer.prompt(list).then((r) => {
+    if (r.OperationType == 'dist') {
+      removeDist()
+    } else if (r.OperationType == 'dist-electron') {
+      removeDistElectron()
+    }
+  })
+}
+
+const removeDist = () => {
   const distPath = path.resolve(__dirname, '../dist')
+
   if (!isDistExist(distPath)) {
     return exitProcess('当前路径下不存在 `dist` 文件夹，3秒后自动退出', 'fail')
   }
@@ -67,7 +93,7 @@ const start = () => {
     return exitProcess('`dist` 文件夹为空，3秒后自动退出', 'succeed')
   }
 
-  const removeList = [
+  const removeListDist = [
     {
       type: 'checkbox',
       message: '选择要删除的文件夹:',
@@ -76,7 +102,7 @@ const start = () => {
     },
   ]
 
-  inquirer.prompt(removeList).then(async (answers) => {
+  inquirer.prompt(removeListDist).then(async (answers) => {
     if (answers.removeDists.length == 0) {
       return exitProcess('未选择任何文件夹，3秒后自动退出', 'fail')
     }
@@ -96,6 +122,49 @@ const start = () => {
 
     answers.removeDists.forEach((p) => {
       const removePath = path.resolve(__dirname, `../dist/${p}`)
+      DeleteFolder(removePath)
+    })
+    return exitProcess('删除成功，3秒后自动退出', 'succeed')
+  })
+}
+const removeDistElectron = () => {
+  const distElectron = path.resolve(__dirname, '../dist-electron')
+
+  const dist_electrons = forEachDist(distElectron)
+  if (!Array.isArray(dist_electrons)) {
+    return exitProcess('`dist-electron` 文件夹为空，3秒后自动退出', 'succeed')
+  }
+  const removeListDistElectron = [
+    {
+      type: 'checkbox',
+      message: '选择要删除的文件夹:',
+      name: 'removeDistsElectron',
+      choices: dist_electrons.map((item) => ({ name: item })),
+    },
+  ]
+  if (!isDistExist(distElectron)) {
+    return exitProcess('当前路径下不存在 `dist-electron` 文件夹，3秒后自动退出', 'fail')
+  }
+  inquirer.prompt(removeListDistElectron).then(async (answers) => {
+    if (answers.removeDistsElectron.length == 0) {
+      return exitProcess('未选择任何文件夹，3秒后自动退出', 'fail')
+    }
+
+    let r = await inquirer.prompt([
+      {
+        type: 'confirm',
+        message: `将要删除以下文件夹\n ${answers.removeDistsElectron} \n 是否继续？`,
+        name: 'isDelete',
+        default: true,
+      },
+    ])
+
+    if (!r.isDelete) {
+      return exitProcess('已取消删除，3秒后自动退出')
+    }
+
+    answers.removeDistsElectron.forEach((p) => {
+      const removePath = path.resolve(__dirname, `../dist-electron/${p}`)
       DeleteFolder(removePath)
     })
     return exitProcess('删除成功，3秒后自动退出', 'succeed')
